// src/components/calculator/CalculatorButtonTheme.d.ts
export enum CalculatorButtonTheme  {
    Light = "bg-stone-50 hover:bg-stone-200 text-stone-800",
    Orange = "bg-amber-600 hover:bg-amber-700 text-stone-100",
    Gray = "bg-stone-800 hover:bg-stone-900 text-stone-50",
    Red = "bg-red-600 hover:bg-red-700 text-stone-50",
    DarkRed = "bg-red-800 hover:bg-red-900 text-stone-50",
    Blue = "bg-blue-600 hover:bg-blue-700 text-stone-50",
    DarkBlue = "bg-blue-800 hover:bg-blue-900 text-stone-50",
    Yellow = "bg-yellow-500 hover:bg-yellow-600 text-stone-100",
    Pink = "bg-pink-600 hover:bg-pink-700 text-stone-100",
    Rose = "bg-rose-600 hover:bg-rose-700 text-stone-100",
    DarkRose = "bg-rose-800 hover:bg-rose-900 text-rose-100",
}